+++ lib/src/codec.dart
@@ -4,7 +4,7 @@ import 'types/base/bson_object.dart';
 class Codec {
   static BsonBinary serialize(var object, SerializationParameters parms) {
     var bsonObject = BsonObject.from(object, parms);
-    var buffer = BsonBinary(bsonObject.byteLength());
+    var buffer = BsonBinary(bsonObject.byteLength);
     bsonObject.packValue(buffer);
     return buffer;
   }