+++ lib/src/types/base/bson_container.dart
@@ -2,6 +2,9 @@ import '../../../bson.dart';
 import 'bson_object.dart';
 
 abstract class BsonContainer extends BsonObject {
+  /// Length of the data elements
+  int get dataSize;
+
   @Deprecated('No more used')
   static int entrySize(String? name, value, SerializationParameters parms) {
     var size = 1;
@@ -9,6 +12,6 @@ abstract class BsonContainer extends BsonObject {
       size += Statics.getKeyUtf8(name).length + 1;
     }
     // TODO Optimize here
-    return size + BsonObject.from(value, parms).byteLength();
+    return size + BsonObject.from(value, parms).byteLength;
   }
 }