+++ lib/src/types/bson_binary.dart
@@ -324,7 +324,7 @@ class BsonBinary extends BsonObject {
   }
 
   @override
-  int byteLength() => byteList.length + 4 + 1;
+  int get byteLength => byteList.length + 4 + 1;
   bool atEnd() => offset == byteList.length;
   void rewind() => offset = 0;