+++ lib/src/types/bson_string.dart
@@ -35,7 +35,7 @@ class BsonString extends BsonObject {
   @override
   dynamic get value => data;
   @override
-  int byteLength() => utfData.length + 1 + 4;
+  int get byteLength => utfData.length + 1 + 4;
   @override
   int get typeByte => bsonDataString;
   @override
@@ -50,41 +50,6 @@ class BsonString extends BsonObject {
   @override
   eJson({bool relaxed = false}) => data;
 }
-
-/* class BsonCode extends BsonString {
-  BsonCode(super.dataValue) : super();
-  BsonCode.fromBuffer(super.buffer) : super.fromBuffer();
-  BsonCode.fromEJson(Map<String, dynamic> eJsonMap)
-      : super(extractEJson(eJsonMap));
-
-  static String extractData(BsonBinary buffer) =>
-      BsonString.extractData(buffer);
-
-  static String extractEJson(Map<String, dynamic> ejsonMap) {
-    var entry = ejsonMap.entries.first;
-    if (entry.key != type$code) {
-      throw ArgumentError(
-          'The received Map is not a avalid EJson Code representation');
-    }
-    if (entry.value is String) {
-      return entry.value;
-    }
-
-    throw ArgumentError(
-        'The received Map is not a valid EJson String representation');
-  }
-
-  @override
-  BsonCode get value => this;
-
-  @override
-  int get typeByte => bsonDataCode;
-  @override
-  String toString() => "BsonCode('$data')";
-
-  @override
-  eJson({bool relaxed = false}) => {type$code: data};
-} */
 
 class BsonCString extends BsonString {
   BsonCString(super.data, {this.useKeyCash = false}) : super();
@@ -117,7 +82,7 @@ class BsonCString extends BsonString {
   }
 
   @override
-  int byteLength() => utfData.length + 1;
+  int get byteLength => utfData.length + 1;
 
   @override
   void packValue(BsonBinary buffer) {